backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.3.0
pydantic>=2.4.0
python-decouple>=3.8
python-multipart>=0.0.6
